inversion counting 해결 - 결과값을 int가 아닌 long long에 저장해서 해결

+++ SW 알고리즘 역량강화/Inversion Counting/Inversion Counting/main.cpp
@@ -38,7 +38,7 @@ constexpr size_t MAX_N = 100001;
 using namespace std;
 
 int arr[MAX_N], buffer[MAX_N];
-int cnt;
+unsigned long long cnt;
 
 void merge(int* const begin, int* const mid, int* const end) {
 	int* begin1 = begin;
